KIRO2022

Documentation for KIRO2022.

Parsing functions

This package exports the three following parsing functions:

read_instance: read an instance file from its path and returns a corresponding Instance object.

read_solution: write a Solution object to a solution format json file.

write_solution: read a solution file into a Solution object.

Data structures

Instance

Methods nb_tasks, nb_operators, nb_jobs, nb_machines can be used to query instance parameters.

Solution

Note: you may want to implement your own custom data structures that contain additional information that may be used by your algorithms.

Evaluation utilities

is_feasible: function that tests if a solution is feasible for a given instance

cost: function that computes the cost of a solution for given instance

Preparing submission

prepare_submission: this function take as input a solver function and automatically produces a solution file for each instance file, ready for submission.

API

KIRO2022.Instance

KIRO2022.Solution

KIRO2022.cost

KIRO2022.is_feasible

KIRO2022.job_cost

KIRO2022.prepare_submission

KIRO2022.read_instance

KIRO2022.read_solution

KIRO2022.write_solution

KIRO2022.Instance — Type

Instance

Attributes

nb_jobs

nb_tasks

nb_machines

unit_penalty: penalty when a job is late

tardiness: delay penalty for jobs

jobs: job list, in index order

tasks: task list, in index order

KIRO2022.Solution — Type

Solution

Attributes

starts: starting date of each task.

machines: machines for each task.

operators: operator for each task.

KIRO2022.cost — Method

cost(solution::Solution, instance::Instance)

Compute the objective value of solution for given instance.

KIRO2022.is_feasible — Method

is_feasible(solution::Solution, instance::Instance; verbose=true)

Check if solution is feasible for instance. Prints some warnings when solution is infeasible and verbose is true.

KIRO2022.job_cost — Method

job_cost(job_index::Int, solution::Solution, instance::Instance)

Compute the value of job job_index in the objective for solution in instance.

KIRO2022.prepare_submission — Method

prepare_submission(solver, instances_folder, solutions_folder, group)

Read instances from the instance_folder, use solver to generate solutions, and then write these solutions in the solutions_folder with the right format and group number.

The solver should be a function with one Instance argument, and return a (feasible) Solution.

KIRO2022.read_instance — Method

read_instance(path::String)

Read instance from json file path.

KIRO2022.read_solution — Method

read_solution(path::String)

Read solution from json file path.

KIRO2022.write_solution — Method

write_solution(solution::Solution, path::String)

Write solution to file path with json format.